Add files via upload
Fill this in using the comments within as a guide. 

The document you are creating is the bare bones of what a html document almost always consists of, the main aim here is to write your own html document so you know HOW and WHY it is structured in this way.

ONLY use the comments provided, if you get stuck, or can't even begin to fill it in, don't worry, go to solo learn (or any other resource) and refresh your memory then come back and try again, good luck! 

- Mike